+++ CHANGELOG.md
@@ -1,1 +1,6 @@
+## New Features
+- Added `IngredientStack` - A class designed to help in recipe serialization for machines that take non-one amounts of inputs of both items and fluids.
+
+## Bug Fixes
 - Fixed connected texture baked model not ported correctly
+- Fixed network handler having private access in event subscribers